use sync file reading and move it to `sany` file
