+++ src/commands/parseModule.ts
@@ -51,28 +51,6 @@ async function transpilePlusCal(fileUri: vscode.Uri): Promise<DCollection> {
     return stdoutParser.readAll();
 }
 
-async function adaptMonolithMessages(sanyResult : SanyData) {
-    const invertedModulePaths = new Map(Array.from(sanyResult.modulePaths, (i) => i.reverse() as [string, string]));
-    for (const [filePath, monolithFilePath] of sanyResult.filePathToMonolithFilePath) {
-        const text = (await vscode.workspace.openTextDocument(monolithFilePath)).getText();
-        const specName = invertedModulePaths.get(filePath);
-        text.split('\n').forEach(function (line, number) {
-            if (new RegExp(`-----*\\s*MODULE\\s+${specName}\\s*----*`).exec(line)) {
-                sanyResult.dCollection.getMessages().filter(m => m.filePath === filePath).forEach(message => {
-                    const oldRange = message.diagnostic.range;
-                    // Remove message so it does not appear duplicated in the output.
-                    sanyResult.dCollection.removeMessage(message);
-                    sanyResult.dCollection.addMessage(
-                        monolithFilePath, 
-                        new vscode.Range(oldRange.start.line + number, oldRange.start.character, oldRange.end.line + number, oldRange.end.character),
-                        message.diagnostic.message,
-                        message.diagnostic.severity);
-                })
-            } 
-        });
-    }
-}
-
 /**
  * Parses the resulting TLA+ spec.
  */
@@ -80,9 +58,5 @@ async function parseSpec(fileUri: vscode.Uri): Promise<SanyData> {
     const procInfo = await runSany(fileUri.fsPath);
     sanyOutChannel.bindTo(procInfo);
     const stdoutParser = new SanyStdoutParser(procInfo.process.stdout);
-    const sanyResult = await stdoutParser.readAll();
-    // Adapt monolith error locations.
-    // It modifies the Sany result adding the module offset for in the monolith spec.
-    await adaptMonolithMessages(sanyResult);
-    return sanyResult;
+    return stdoutParser.readAll();
 }
